Is valid parenthesis - stack based solution.
A single stack stores the open parenenthesis. Each time a closing parenthesis is visited, we're popping the stack to see if there's a matching opening parenthesis in the stack. If yes, we continue. If not, we return false.

In the end, we're checking if there are any chars left in the stack. If empty -> the input was valid.