Minimum Window Substring - find a minimum substring in string s that contains all the chars from the string t. Solution: two pointers traversing the s string: left and right. Maintaining a map of chars and the count left from the t string.
When visiting a new char by moving the right pointer - decrementing its count in the map. If at the same time, all the chars from t are available - saving the best result if it's the best one.

When moving the left pointer - incrementing the counter for the removed char from the map.

Traversing until the left pointer hits the end of the s array.