Is isomorphic string - Given two strings s and t, determine if they are isomorphic.
Two strings s and t are isomorphic if the characters in s can be replaced to get t.

All occurrences of a character must be replaced with another character while preserving the order of characters. No two characters may map to the same character, but a character may map to itself.

Solution: keeping a map to store `s` to `t` transformations. Keeping a set to make sure that a `t` character is not mapped twice. After checking the legnths are equal, traversing the indices:
1. If `s` char is already in the transformations map- continue
2. If the `s` char is not in the transformation map and the `t` char has not been seen before: add to the transformation map and visited ts
3. Else, i.e. this `s` char has already been mapped and to another `t` - return false.
Once gone through all the indices, return true.