Given an m x n board of characters and a list of strings words, return all words on the board.
Each word must be constructed from letters of sequentially adjacent cells, where adjacent cells are horizontally or vertically neighboring. The same letter cell may not be used more than once in a word.

Solution: keeping a trie that contains all the words to search for. Keeping a list of words already found. Having a visit() method that takes the current TrieNode, n, m, and path that lead to this node. This method
1. Checks if it was already visited or n, m are outside od range. If yes - return
2. Checks the current char and if it's in the current TrieNode. If yes, it checks if the next TrieNode is a "full word" node. It also visits() all the adjacent nodes, providing the nextNode as a param.
3. It visits() all the adjacent nodes, providing the trie root node as the current node