Given the root of a complete binary tree, return the number of the nodes in the tree.
According to Wikipedia, every level, except possibly the last, is completely filled in a complete binary tree, and all nodes in the last level are as far left as possible. It can have between 1 and 2h nodes inclusive at the last level h.

Design an algorithm that runs in less than O(n) time complexity.

Solution: naively, we can do it recursively, but then we'll have O(n). Better: notice, that the last node is always going to be on the last level of the tree and there will be no nodes to it right. Becase of that, the problem can be viewed as binary search that traverses the last level of the tree until it finds a node that has no node on the right.

1. The entire algo begins with checking the height of the tree.
2. Then we do binary search: