Finding subarray that sums to a target value in an array of positive integers.
Solution: sliding window with l and r pointers. Knowing that all items are positive, adding item to the subarray is always going to increase curent sum. So when the current subarray is too small - move the r pointer to the right. If the sum is to large - move the l pointer to the left.