Max Subset Sum No Adjacent - find max sum of items in a non-negative array of non-adjacent items.
Solution: notice that it can be taken from left to right. We could build an array of items, calculating what's the highest current sum for a specific item, looking at the previous items. Then notice, we don't need the entire array, we just need to keep the i-1 and i-2 values. What's important, the i-2 value doesn't have to be always replaced with i-1. Instead it could stay as it is, if it's of a higher value.